Close beta v 1.1

+++ app/src/main/java/main/work/braintrainercompose/game/domain/models/GameResults.kt
@@ -3,5 +3,5 @@ package main.work.braintrainercompose.game.domain.models
 data class GameResults(
     val score: String = "",
     val time: String? = "00:00",
-    val accuracy: Float = 100F
+    val accuracy: Int = 100
 )
